+++ tests/org.palladiosimulator.simexp.dsl.ea.optimizer.tests/src/main/java/org/palladiosimulator/simexp/dsl/ea/optimizer/pareto/ParetoEvolutionStatisticsTest.java
@@ -3,6 +3,8 @@ package org.palladiosimulator.simexp.dsl.ea.optimizer.pareto;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.MockitoAnnotations.initMocks;
 
+import java.math.BigInteger;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -19,19 +21,19 @@ public class ParetoEvolutionStatisticsTest {
     public void setUp() throws Exception {
         initMocks(this);
 
-        statistics = new ParetoEvolutionStatistics<>(fitnessFunction, 1l);
+        statistics = new ParetoEvolutionStatistics<>(fitnessFunction, BigInteger.valueOf(1l));
     }
 
     @Test
     public void testToRoundedString1() {
-        String actualRoundedString = statistics.toRoundedString(1, 100);
+        String actualRoundedString = statistics.toRoundedString(1, BigInteger.valueOf(100));
 
         assertThat(actualRoundedString).isEqualTo("1");
     }
 
     @Test
     public void testToRoundedString100() {
-        String actualRoundedString = statistics.toRoundedString(100, 100);
+        String actualRoundedString = statistics.toRoundedString(100, BigInteger.valueOf(100));
 
         assertThat(actualRoundedString).isEqualTo("100");
     }
